requirements.txt
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
fonttools==4.38.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
plotly==5.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tenacity==8.1.0
Werkzeug==2.2.2
zipp==3.11.0
